First step to make SEP c++ and thread-safe
